+++ test/quantum/job_broadcaster_test.exs
@@ -6,8 +6,10 @@ defmodule Quantum.JobBroadcasterTest do
   alias Quantum.JobBroadcaster
   alias Quantum.TestConsumer
   alias Quantum.Job
+  alias Quantum.Storage.Test, as: TestStorage
 
   import ExUnit.CaptureLog
+  import Quantum.CaptureLogExtend
 
   doctest JobBroadcaster
 
@@ -18,6 +20,10 @@ defmodule Quantum.JobBroadcasterTest do
   end
 
   setup tags do
+    if tags[:listen_storage] do
+      Process.put(:test_pid, self())
+    end
+
     active_job = TestScheduler.new_job()
     inactive_job = Job.set_state(TestScheduler.new_job(), :inactive)
 
@@ -36,8 +42,21 @@ defmodule Quantum.JobBroadcasterTest do
           []
       end
 
-    {:ok, broadcaster} = start_supervised({JobBroadcaster, {__MODULE__, init_jobs, true}})
-    {:ok, _consumer} = start_supervised({TestConsumer, [broadcaster, self()]})
+    broadcaster =
+      if tags[:manual_dispatch] do
+        nil
+      else
+        {{:ok, broadcaster}, _} =
+          capture_log_with_return(fn ->
+            start_supervised(
+              {JobBroadcaster, {__MODULE__, init_jobs, TestStorage, TestScheduler, true}}
+            )
+          end)
+
+        {:ok, _consumer} = start_supervised({TestConsumer, [broadcaster, self()]})
+
+        broadcaster
+      end
 
     {
       :ok,
@@ -56,14 +75,42 @@ defmodule Quantum.JobBroadcasterTest do
       refute_receive {:received, {:add, ^inactive_job}}
       assert_receive {:received, {:add, ^active_job}}
     end
+
+    @tag manual_dispatch: true
+    test "storage jobs", %{active_job: active_job, inactive_job: inactive_job} do
+      capture_log(fn ->
+        defmodule FullStorage do
+          @moduledoc false
+
+          use Quantum.Storage.Test
+
+          def jobs(_),
+            do: [
+              TestScheduler.new_job(),
+              Job.set_state(TestScheduler.new_job(), :inactive)
+            ]
+        end
+
+        {:ok, broadcaster} =
+          start_supervised({JobBroadcaster, {__MODULE__, [], FullStorage, TestScheduler, true}})
+
+        {:ok, _consumer} = start_supervised({TestConsumer, [broadcaster, self()]})
+
+        assert_receive {:received, {:add, _}}
+        refute_receive {:received, {:add, _}}
+      end)
+    end
   end
 
   describe "add" do
+    @tag listen_storage: true
     test "active", %{broadcaster: broadcaster, active_job: active_job} do
       assert capture_log(fn ->
                TestScheduler.add_job(broadcaster, active_job)
 
                assert_receive {:received, {:add, ^active_job}}
+
+               assert_receive {:add_job, {TestScheduler, ^active_job}, _}
              end) =~ "Adding job #Reference"
     end
 
@@ -74,7 +121,9 @@ defmodule Quantum.JobBroadcasterTest do
                :ok = stop_supervised(TestConsumer)
 
                {:ok, broadcaster} =
-                 start_supervised({JobBroadcaster, {__MODULE__, init_jobs, false}})
+                 start_supervised(
+                   {JobBroadcaster, {__MODULE__, init_jobs, TestStorage, TestScheduler, false}}
+                 )
 
                {:ok, _consumer} = start_supervised({TestConsumer, [broadcaster, self()]})
 
@@ -84,17 +133,20 @@ defmodule Quantum.JobBroadcasterTest do
              end) =~ "Adding job #Reference"
     end
 
+    @tag listen_storage: true
     test "inactive", %{broadcaster: broadcaster, inactive_job: inactive_job} do
       capture_log(fn ->
         TestScheduler.add_job(broadcaster, inactive_job)
 
         refute_receive {:received, {:add, _}}
+
+        assert_receive {:add_job, {TestScheduler, ^inactive_job}, _}
       end)
     end
   end
 
   describe "delete" do
-    @tag jobs: :active
+    @tag jobs: :active, listen_storage: true
     test "active", %{broadcaster: broadcaster, active_job: active_job} do
       active_job_name = active_job.name
 
@@ -103,26 +155,35 @@ defmodule Quantum.JobBroadcasterTest do
 
         assert_receive {:received, {:remove, ^active_job_name}}
 
+        assert_receive {:delete_job, {TestScheduler, ^active_job_name}, _}
+
         refute Enum.any?(TestScheduler.jobs(broadcaster), fn {key, _} ->
                  key == active_job_name
                end)
       end)
     end
 
+    @tag listen_storage: true
     test "missing", %{broadcaster: broadcaster} do
       capture_log(fn ->
         TestScheduler.delete_job(broadcaster, make_ref())
 
         refute_receive {:received, {:remove, _}}
+
+        refute_receive {:delete_job, {TestScheduler, _}, _}
       end)
     end
 
-    @tag jobs: :inactive
+    @tag jobs: :inactive, listen_storage: true
     test "inactive", %{broadcaster: broadcaster, inactive_job: inactive_job} do
       capture_log(fn ->
+        inactive_job_name = inactive_job.name
+
         TestScheduler.delete_job(broadcaster, inactive_job.name)
 
         refute_receive {:received, {:remove, _}}
+
+        assert_receive {:delete_job, {TestScheduler, ^inactive_job_name}, _}
 
         refute Enum.any?(TestScheduler.jobs(broadcaster), fn {key, _} ->
                  key == inactive_job.name
@@ -132,7 +193,7 @@ defmodule Quantum.JobBroadcasterTest do
   end
 
   describe "change_state" do
-    @tag jobs: :active
+    @tag jobs: :active, listen_storage: true
     test "active => inactive", %{broadcaster: broadcaster, active_job: active_job} do
       active_job_name = active_job.name
 
@@ -140,10 +201,12 @@ defmodule Quantum.JobBroadcasterTest do
         TestScheduler.deactivate_job(broadcaster, active_job.name)
 
         assert_receive {:received, {:remove, ^active_job_name}}
+
+        assert_receive {:update_job_state, {TestScheduler, _, _}, _}
       end)
     end
 
-    @tag jobs: :inactive
+    @tag jobs: :inactive, listen_storage: true
     test "inactive => active", %{broadcaster: broadcaster, inactive_job: inactive_job} do
       capture_log(fn ->
         TestScheduler.activate_job(broadcaster, inactive_job.name)
@@ -151,10 +214,12 @@ defmodule Quantum.JobBroadcasterTest do
         active_job = Job.set_state(inactive_job, :active)
 
         assert_receive {:received, {:add, ^active_job}}
+
+        assert_receive {:update_job_state, {TestScheduler, _, _}, _}
       end)
     end
 
-    @tag jobs: :active
+    @tag jobs: :active, listen_storage: true
     test "active => active", %{broadcaster: broadcaster, active_job: active_job} do
       # Initial
       assert_receive {:received, {:add, ^active_job}}
@@ -163,10 +228,12 @@ defmodule Quantum.JobBroadcasterTest do
         TestScheduler.activate_job(broadcaster, active_job.name)
 
         refute_receive {:received, {:add, ^active_job}}
+
+        refute_receive {:update_job_state, {TestScheduler, _, _}, _}
       end)
     end
 
-    @tag jobs: :inactive
+    @tag jobs: :inactive, listen_storage: true
     test "inactive => inactive", %{broadcaster: broadcaster, inactive_job: inactive_job} do
       inactive_job_name = inactive_job.name
 
@@ -174,9 +241,12 @@ defmodule Quantum.JobBroadcasterTest do
         TestScheduler.deactivate_job(broadcaster, inactive_job.name)
 
         refute_receive {:received, {:remove, ^inactive_job_name}}
+
+        refute_receive {:update_job_state, {TestScheduler, _, _}, _}
       end)
     end
 
+    @tag listen_storage: true
     test "missing", %{broadcaster: broadcaster} do
       capture_log(fn ->
         TestScheduler.deactivate_job(broadcaster, make_ref())
@@ -184,12 +254,13 @@ defmodule Quantum.JobBroadcasterTest do
 
         refute_receive {:received, {:remove, _}}
         refute_receive {:received, {:add, _}}
+        refute_receive {:update_job_state, {TestScheduler, _, _}, _}
       end)
     end
   end
 
   describe "delete_all" do
-    @tag jobs: :both
+    @tag jobs: :both, listen_storage: true
     test "only active jobs", %{
       broadcaster: broadcaster,
       active_job: active_job,
@@ -203,6 +274,8 @@ defmodule Quantum.JobBroadcasterTest do
 
         refute_receive {:received, {:remove, ^inactive_job_name}}
         assert_receive {:received, {:remove, ^active_job_name}}
+
+        assert_receive {:purge, TestScheduler, _}
       end)
     end
   end